pyproject.toml
[project]
name = "OmniSafeBench-MM"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "backoff>=2.2.1",
    "decord>=0.6.0",
    "diffusers>=0.35.2",
    "einops>=0.8.1",
    "iopath>=0.1.10",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "omegaconf>=2.3.0",
    "openai>=2.9.0",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "peft==0.17.1",
    "pillow>=12.0.0",
    "pip>=25.3",
    "protobuf>=6.33.2",
    "pyyaml>=6.0.3",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "sentence-transformers",
    "sentencepiece>=0.2.1",
    "spacy>=3.8.11",
    "textaugment>=2.0.0",
    "textblob==0.15.3",
    "timm>=1.0.22",
    "torch==2.6.0+cu118",
    "torchvision==0.21.0+cu118",
    "tqdm>=4.67.1",
    "transformers==4.45",
    "visual-genome>=1.1.1",
    "wandb>=0.23.1",
    "webdataset>=1.0.2",
]
[tool.pytest.ini_options]
testpaths = ["tests"]


[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
